colorbook/requirements.txt
Flask>=3.0,<4.0
Werkzeug>=3.0,<4.0
gunicorn>=22.0
openai>=1.40
Pillow>=10.0
